feat(gfx_d3d): land upstream r_reflection_probe into POSIX/Switch build
Wire src/gfx_d3d/r_reflection_probe.cpp (546 lines) into both
backends — disk-side reflection-probe load / color-correction LUT
load and the BSP-wide reflection generator entry
(R_BspGenerateReflections / R_GenerateReflectionImages /
R_LoadColorCorrectionData).

Source fixes for -Werror clean build:
  - KISAKHACK uintptr_t bridge on the
    `(unsigned __int8 *)Hunk_AllocateTempMemoryHigh(...)` 32-bit
    truncation in R_LoadColorCorrectionData.

Dup removal in posix_backbone_stubs:
  - Comment out R_BspGenerateReflections and R_GenerateReflectionImages
    stubs now provided by upstream.

Satellite stubs:
  - Image_Alloc / Image_GenerateCube (live in r_image; nullptr/no-op
    until r_image lands).
  - R_CreateReflectionRawDataFromCubemapShot (lives in r_screenshot).

+++ src/gfx_d3d/r_reflection_probe.cpp
@@ -246,7 +246,7 @@ void R_LoadColorCorrectionData()
     if (fileSize >= 0)
     {
         Hunk_CheckTempMemoryHighClear();
-        filebuf = (unsigned __int8 *)Hunk_AllocateTempMemoryHigh(fileSize + 1, "R_LoadColorCorrectionData");
+        filebuf = (unsigned __int8 *)(uintptr_t)Hunk_AllocateTempMemoryHigh(fileSize + 1, "R_LoadColorCorrectionData");
         FS_Read(filebuf, fileSize, f);
         FS_FCloseFile(f);
         filebuf[fileSize] = 0;
